requirements.txt
django
djangorestframework
pillow
django-phonenumber-field[phonenumberslite]
djangorestframework-simplejwt